너무무서워요
내 2주어디감

+++ Assets/22_Song/SaveLoad/TileSaveData.cs
@@ -6,7 +6,7 @@ public class TileSaveData
 {
 	public int index;
 
-	public int tileWeatValue;
+	public float tileWeatValue;
 
 	public int tileExp;
 	public float tileHp;
